Drop FragmentType + load tagging: the Mma is the sole tensor-core marker
The operand Loads never needed a tag. The Mma already carries the atom kind
and names its A/B operands by SSA value, and that association survives every
pass (staging keeps SSA names; pipelining/register-split duplicate the whole
Load+Load+Mma cell together). So the lowering recovers each operand's role
from the co-located Mma — no tag on the loads, no FragmentType.

- Remove FragmentType; operand Loads stay plain (drop Load.atom/role props).
- 011_lower_atom_cell: emit plain Loads + an Mma (no tagging).
- 005/_mma: recover A/B per reduce site via the Mma's operand names; seed
  fragment SSA names from the Mma (every reduce site shares names).
- Revert the dtype= preservation I'd added to 020/040 for the tag (loads are
  plain again; tile-stage dtype is None anyway).
- Keep the StructuredType / F16x2 split — useful on its own for the packed
  __half2 type.

Kernel output byte-identical; full suite green (1572). Docs refreshed.

+++ deplodock/compiler/pipeline/passes/lowering/tile/020_stage_inputs.py
@@ -711,9 +711,7 @@ def _rewrite_loads(stmt: Stmt, name_rewrites: dict[str, tuple[str, tuple[Expr, .
     def fn(s: Stmt) -> Stmt:
         if isinstance(s, Load) and s.name in name_rewrites:
             smem_name, new_index = name_rewrites[s.name]
-            # Preserve ``dtype`` so a fragment Load (dtype is a FragmentType)
-            # still lowers to ldmatrix at kernel/005_lower_atom_tile.
-            return Load(name=s.name, input=smem_name, index=new_index, dtype=s.dtype)
+            return Load(name=s.name, input=smem_name, index=new_index)
         return s
 
     return Body((stmt,)).map(fn)[0]
